fix(GiftCardChargeEventListener.kt): 입금 금액을 정수로 변환하여 전송 feat(SmsService.kt): SMS 전송 시 메세지 전송 로그 추가 chore(StoreScheduler.kt): 주기적 스케줄러 작업 제거

+++ src/main/kotlin/com/cuberlabs/cuperpinserver/domain/giftcardcharge/service/event/listner/GiftCardChargeEventListener.kt
@@ -16,6 +16,7 @@ class GiftCardChargeEventListener(
 ) {
     @Value("\${sms.admin-number}")
     var adminNumber: String = ""
+
     @EventListener
     fun handlerChargeCompleteEvent(event: GiftCardChargeCompleteEvent) {
         // 관리자에게 입금 요청
@@ -25,7 +26,7 @@ class GiftCardChargeEventListener(
                 SmsMessages.depositRequest(
                     bank = bank.toString(),
                     accountNumber = accountNumber,
-                    amount = depositAmount.toString()
+                    amount = depositAmount.toInt().toString()
                 )
             )
         }
